added GP option
