+++ elfi/methods/parameter_inference.py
@@ -23,7 +23,7 @@ from elfi.methods.results import BolfiSample, OptimizationResult, Sample, SmcSam
 from elfi.methods.utils import (GMDistribution, ModelPrior, arr2d_to_batch,
                                 batch_to_arr2d, ceil_to_batch_size, weighted_var, create_deterministic_generator,
                                 create_output_function, OptimisationProblem,
-                                collect_solutions, compute_ess, GPTrainer)
+                                collect_solutions, compute_ess)
 from elfi.model.elfi_model import ComputationContext, ElfiModel, NodeReference
 from elfi.utils import is_array
 from elfi.visualization.visualization import ProgressBar
@@ -1355,6 +1355,404 @@ class BOLFI(BayesianOptimization):
             seed=self.seed)
 
 
+class BoDetereministic:
+    """Bayesian Optimization of an unknown target function."""
+
+    def __init__(self,
+                 det_func,
+                 prior,
+                 parameter_names,
+                 n_evidence,
+                 target_name=None,
+                 bounds=None,
+                 initial_evidence=None,
+                 update_interval=10,
+                 target_model=None,
+                 acquisition_method=None,
+                 acq_noise_var=0,
+                 exploration_rate=10,
+                 batch_size=1,
+                 async_acq=False,
+                 seed=None,
+                 **kwargs):
+        """Initialize Bayesian optimization.
+
+        Parameters
+        ----------
+        det_func : ElfiModel or NodeReference
+        target_name : str or NodeReference
+            Only needed if model is an ElfiModel
+        bounds : dict, optional
+            The region where to estimate the posterior for each parameter in
+            model.parameters: dict('parameter_name':(lower, upper), ... )`. Not used if
+            custom target_model is given.
+        initial_evidence : int, dict, optional
+            Number of initial evidence or a precomputed batch dict containing parameter
+            and discrepancy values. Default value depends on the dimensionality.
+        update_interval : int, optional
+            How often to update the GP hyperparameters of the target_model
+        target_model : GPyRegression, optional
+        acquisition_method : Acquisition, optional
+            Method of acquiring evidence points. Defaults to LCBSC.
+        acq_noise_var : float or np.array, optional
+            Variance(s) of the noise added in the default LCBSC acquisition method.
+            If an array, should be 1d specifying the variance for each dimension.
+        exploration_rate : float, optional
+            Exploration rate of the acquisition method
+        batch_size : int, optional
+            Elfi batch size. Defaults to 1.
+        batches_per_acquisition : int, optional
+            How many batches will be requested from the acquisition function at one go.
+            Defaults to max_parallel_batches.
+        async_acq : bool, optional
+            Allow acquisitions to be made asynchronously, i.e. do not wait for all the
+            results from the previous acquisition before making the next. This can be more
+            efficient with a large amount of workers (e.g. in cluster environments) but
+            forgoes the guarantee for the exactly same result with the same initial
+            conditions (e.g. the seed). Default False.
+        **kwargs
+
+        """
+        self.det_func = det_func
+
+        self.prior = prior
+
+        self.bounds = bounds
+        self.batch_size = batch_size
+        self.parameter_names = parameter_names
+        self.seed = seed
+
+        self.target_name = target_name
+        self.target_model = target_model
+
+        n_precomputed = 0
+        n_initial, precomputed = self._resolve_initial_evidence(initial_evidence)
+        if precomputed is not None:
+            params = batch_to_arr2d(precomputed, self.parameter_names)
+            n_precomputed = len(params)
+            self.target_model.update(params, precomputed[target_name])
+
+        self.batches_per_acquisition = 1
+        self.acquisition_method = acquisition_method or LCBSC(self.target_model,
+                                                              prior=self.prior,
+                                                              noise_var=acq_noise_var,
+                                                              exploration_rate=exploration_rate,
+                                                              seed=self.seed)
+
+        self.n_initial_evidence = n_initial
+        self.n_precomputed_evidence = n_precomputed
+        self.update_interval = update_interval
+        self.async_acq = async_acq
+
+        self.state = {}
+        self.state['n_evidence'] = self.n_precomputed_evidence
+        self.state['last_GP_update'] = self.n_initial_evidence
+        self.state['acquisition'] = []
+        self.state['n_sim'] = 0
+        self.state['n_batches'] = 0
+
+        self.set_objective(n_evidence)
+
+    def _resolve_initial_evidence(self, initial_evidence):
+        # Some sensibility limit for starting GP regression
+        precomputed = None
+        n_required = max(10, 2 ** self.target_model.input_dim + 1)
+        n_required = ceil_to_batch_size(n_required, self.batch_size)
+
+        if initial_evidence is None:
+            n_initial_evidence = n_required
+        elif isinstance(initial_evidence, (int, np.int, float)):
+            n_initial_evidence = int(initial_evidence)
+        else:
+            precomputed = initial_evidence
+            n_initial_evidence = len(precomputed[self.target_name])
+
+        if n_initial_evidence < 0:
+            raise ValueError('Number of initial evidence must be positive or zero '
+                             '(was {})'.format(initial_evidence))
+        elif n_initial_evidence < n_required:
+            logger.warning('We recommend having at least {} initialization points for '
+                           'the initialization (now {})'.format(n_required, n_initial_evidence))
+
+        if precomputed is None and (n_initial_evidence % self.batch_size != 0):
+            logger.warning('Number of initial_evidence %d is not divisible by '
+                           'batch_size %d. Rounding it up...' % (n_initial_evidence,
+                                                                 self.batch_size))
+            n_initial_evidence = ceil_to_batch_size(n_initial_evidence, self.batch_size)
+
+        return n_initial_evidence, precomputed
+
+    @property
+    def n_evidence(self):
+        """Return the number of acquired evidence points."""
+        return self.state.get('n_evidence', 0)
+
+    @property
+    def acq_batch_size(self):
+        """Return the total number of acquisition per iteration."""
+        return self.batch_size * self.batches_per_acquisition
+
+    def set_objective(self, n_evidence=None):
+        """Set objective for inference.
+
+        You can continue BO by giving a larger n_evidence.
+
+        Parameters
+        ----------
+        n_evidence : int
+            Number of total evidence for the GP fitting. This includes any initial
+            evidence.
+
+        """
+        if n_evidence is None:
+            n_evidence = self.objective.get('n_evidence', self.n_evidence)
+
+        if n_evidence < self.n_evidence:
+            logger.warning('Requesting less evidence than there already exists')
+
+        self.objective = {'n_evidence': n_evidence,
+                          'n_sim': n_evidence - self.n_precomputed_evidence}
+
+    def _extract_result_kwargs(self):
+        """Extract common arguments for the ParameterInferenceResult object."""
+        return {
+            'method_name': self.__class__.__name__,
+            'parameter_names': self.parameter_names,
+            'seed': self.seed,
+            'n_sim': self.state['n_sim'],
+            'n_batches': self.state['n_batches']
+        }
+
+    def extract_result(self):
+        """Extract the result from the current state.
+
+        Returns
+        -------
+        OptimizationResult
+
+        """
+        x_min, _ = stochastic_optimization(
+            self.target_model.predict_mean, self.target_model.bounds, seed=self.seed)
+
+        batch_min = arr2d_to_batch(x_min, self.parameter_names)
+        outputs = arr2d_to_batch(self.target_model.X, self.parameter_names)
+        outputs[self.target_name] = self.target_model.Y
+
+        return OptimizationResult(
+            x_min=batch_min, outputs=outputs, **self._extract_result_kwargs())
+
+    def update(self, batch, batch_index):
+        """Update the GP regression model of the target node with a new batch.
+
+        Parameters
+        ----------
+        batch : dict
+            dict with `self.outputs` as keys and the corresponding outputs for the batch
+            as values
+        batch_index : int
+
+        """
+        # super(BayesianOptimization, self).update(batch, batch_index)
+        self.state['n_evidence'] += self.batch_size
+
+        params = batch_to_arr2d(batch, self.parameter_names)
+        self._report_batch(batch_index, params, batch[self.target_name])
+
+        optimize = self._should_optimize()
+        self.target_model.update(params, batch[self.target_name], optimize)
+        if optimize:
+            self.state['last_GP_update'] = self.target_model.n_evidence
+
+    def prepare_new_batch(self, batch_index):
+        """Prepare values for a new batch.
+
+        Parameters
+        ----------
+        batch_index : int
+            next batch_index to be submitted
+
+        Returns
+        -------
+        batch : dict or None
+            Keys should match to node names in the model. These values will override any
+            default values or operations in those nodes.
+
+        """
+        t = self._get_acquisition_index(batch_index)
+
+        # Check if we still should take initial points from the prior
+        if t < 0:
+            return None, None
+
+        # Take the next batch from the acquisition_batch
+        acquisition = self.state['acquisition']
+        if len(acquisition) == 0:
+            acquisition = self.acquisition_method.acquire(self.acq_batch_size, t=t)
+
+        batch = arr2d_to_batch(acquisition[:self.batch_size], self.parameter_names)
+        self.state['acquisition'] = acquisition[self.batch_size:]
+
+        return acquisition, batch
+
+    def _get_acquisition_index(self, batch_index):
+        acq_batch_size = self.batch_size * self.batches_per_acquisition
+        initial_offset = self.n_initial_evidence - self.n_precomputed_evidence
+        starting_sim_index = self.batch_size * batch_index
+
+        t = (starting_sim_index - initial_offset) // acq_batch_size
+        return t
+
+    def fit(self):
+        for ii in range(self.objective["n_sim"]):
+            inp, next_batch = self.prepare_new_batch(ii)
+
+            if inp is None:
+                inp = self.prior.rvs(size=1)
+                if inp.ndim == 1:
+                    inp = np.expand_dims(inp, -1)
+                next_batch = arr2d_to_batch(inp, self.parameter_names)
+
+            y = np.array([self.det_func(np.squeeze(inp, 0))])
+            next_batch[self.target_name] = y
+            self.update(next_batch, ii)
+
+            self.state['n_batches'] += 1
+            self.state['n_sim'] += 1
+
+            toc = timeit.default_timer()
+
+        self.result = self.extract_result()
+
+    # # TODO: use state dict
+    # @property
+    # def _n_submitted_evidence(self):
+    #     return self.batches.total * self.batch_size
+    #
+    # def _allow_submit(self, batch_index):
+    #     if not super(BayesianOptimization, self)._allow_submit(batch_index):
+    #         return False
+    #
+    #     if self.async_acq:
+    #         return True
+    #
+    #     # Allow submitting freely as long we are still submitting initial evidence
+    #     t = self._get_acquisition_index(batch_index)
+    #     if t < 0:
+    #         return True
+    #
+    #     # Do not allow acquisition until previous acquisitions are ready (as well
+    #     # as all initial acquisitions)
+    #     acquisitions_left = len(self.state['acquisition'])
+    #     if acquisitions_left == 0 and self.batches.has_pending:
+    #         return False
+    #
+    #     return True
+
+    def _should_optimize(self):
+        current = self.target_model.n_evidence + self.batch_size
+        next_update = self.state['last_GP_update'] + self.update_interval
+        return current >= self.n_initial_evidence and current >= next_update
+
+    def _report_batch(self, batch_index, params, distances):
+        str = "Received batch {}:\n".format(batch_index)
+        fill = 6 * ' '
+        for i in range(self.batch_size):
+            str += "{}{} at {}\n".format(fill, distances[i].item(), params[i])
+        logger.debug(str)
+
+    def plot_state(self, **options):
+        """Plot the GP surface.
+
+        This feature is still experimental and currently supports only 2D cases.
+        """
+        f = plt.gcf()
+        if len(f.axes) < 2:
+            f, _ = plt.subplots(1, 2, figsize=(13, 6), sharex='row', sharey='row')
+
+        gp = self.target_model
+
+        # Draw the GP surface
+        visin.draw_contour(
+            gp.predict_mean,
+            gp.bounds,
+            self.parameter_names,
+            title='GP target surface',
+            points=gp.X,
+            axes=f.axes[0],
+            **options)
+
+        # Draw the latest acquisitions
+        if options.get('interactive'):
+            point = gp.X[-1, :]
+            if len(gp.X) > 1:
+                f.axes[1].scatter(*point, color='red')
+
+        displays = [gp._gp]
+
+        if options.get('interactive'):
+            from IPython import display
+            displays.insert(
+                0,
+                display.HTML('<span><b>Iteration {}:</b> Acquired {} at {}</span>'.format(
+                    len(gp.Y), gp.Y[-1][0], point)))
+
+        # Update
+        visin._update_interactive(displays, options)
+
+        def acq(x):
+            return self.acquisition_method.evaluate(x, len(gp.X))
+
+        # Draw the acquisition surface
+        visin.draw_contour(
+            acq,
+            gp.bounds,
+            self.parameter_names,
+            title='Acquisition surface',
+            points=None,
+            axes=f.axes[1],
+            **options)
+
+        if options.get('close'):
+            plt.close()
+
+    def plot_discrepancy(self, axes=None, **kwargs):
+        """Plot acquired parameters vs. resulting discrepancy.
+
+        Parameters
+        ----------
+        axes : plt.Axes or arraylike of plt.Axes
+
+        Return
+        ------
+        axes : np.array of plt.Axes
+
+        """
+        return vis.plot_discrepancy(self.target_model, self.parameter_names, axes=axes, **kwargs)
+
+    def plot_gp(self, axes=None, resol=50, const=None, bounds=None, true_params=None, **kwargs):
+        """Plot pairwise relationships as a matrix with parameters vs. discrepancy.
+
+        Parameters
+        ----------
+        axes : matplotlib.axes.Axes, optional
+        resol : int, optional
+            Resolution of the plotted grid.
+        const : np.array, optional
+            Values for parameters in plots where held constant. Defaults to minimum evidence.
+        bounds: list of tuples, optional
+            List of tuples for axis boundaries.
+        true_params : dict, optional
+            Dictionary containing parameter names with corresponding true parameter values.
+
+        Returns
+        -------
+        axes : np.array of plt.Axes
+
+        """
+        return vis.plot_gp(self.target_model, self.parameter_names, axes,
+                           resol, const, bounds, true_params, **kwargs)
+
+
 class ROMC(ParameterInference):
 
     inference_state: Dict
@@ -1514,6 +1912,7 @@ class ROMC(ParameterInference):
 
         # update state
         self.solved = solved
+        self.attempted = attempted
         self.method_state["_has_solved_problems"] = True
 
     def _fit_GP(self, n_evidence, seed=None):
@@ -1529,20 +1928,38 @@ class ROMC(ParameterInference):
 
         gp_trainers = []
         gp_models = []
+        gp_optim_results = []
+        attempted = []
+        solved = []
+        print("### Fitting Gaussian Processes ###")
+        tic = timeit.default_timer()
         for i, func in enumerate(det_funcs):
-
+            progress_bar(i, len(det_funcs), prefix='Progress:', suffix='Complete', length=50)
+            
+            attempted.append(True)
             target_model = GPyRegression(parameter_names=parameter_names, bounds=bounds)
-            trainer = GPTrainer(func, prior, parameter_names, n_evidence, target_name,
-                                bounds=bounds, target_model=target_model)
+            trainer = BoDetereministic(func, prior, parameter_names, n_evidence, target_name,
+                                       bounds=bounds, target_model=target_model, acq_noise_var=0.1)
 
             trainer.fit()
 
             gp_trainers.append(trainer)
             gp_models.append(trainer.target_model.predict_mean)
+            gp_optim_results.append(trainer.result)
+            solved.append(True)
 
+            progress_bar(i+1, len(det_funcs), prefix='Progress:', suffix='Complete', length=50)
 
+        toc = timeit.default_timer()
+        print("Time: %.3f sec" % (toc-tic))
         self.gp_trainers = gp_trainers
+        self.gp_models = gp_models
+        self.gp_optim_results = gp_optim_results
 
+        self.attempted = attempted
+        self.solved = solved
+        self.method_state["_has_solved_problems"] = True
+        self.method_state["_has_fitted_GP"] = True
 
     def _compute_eps(self, quant):
         assert isinstance(quant, float)
@@ -1675,15 +2092,22 @@ class ROMC(ParameterInference):
         print("NOF solutions obtained    : %d " % np.sum(self.solved))
         print("NOF accepted solutions    : %d " % np.sum(self.accepted))
         
-    def solve_problems(self, n1, seed):
+    def solve_problems(self, n1, seed, mode="gradient_based"):
+        assert mode in ["gradient_based", "bayesian_optimization"]
         self._sample_nuisance(n1=n1, seed=seed)
         self._define_optim_problems()
-
-        print("### Solving problems ###")
-        tic = timeit.default_timer()
-        self._solve_optim_problems()
-        toc = timeit.default_timer()
-        print("Time: %.3f sec" % (toc - tic))
+        if mode == "gradient_based":
+            print("### Solving problems ###")
+            tic = timeit.default_timer()
+            self._solve_optim_problems()
+            toc = timeit.default_timer()
+            print("Time: %.3f sec" % (toc - tic))
+        elif mode == "bayesian_optimization":
+            print("### Solving problems ###")
+            tic = timeit.default_timer()
+            self._fit_GP(n_evidence=20, seed=seed)
+            toc = timeit.default_timer()
+            print("Time: %.3f sec" % (toc - tic))
 
     def estimate_regions(self, eps, region_mode):
         # if nothing has been done, stop and print
